[adjust_ww_frequencies] Change WW/EU frequency range
The Medtronic site (URL below) says the frequency for the Guardian is 868.35 and
the USA pump is 916.5

Since the USA frequency is very well optimised, I've tried keeping a similar
size range to the USA frequency, starting close to the 868.35 value like the USA
range starts at 916.5.

I've extended it very slightly since I think the EU frequencies might have a
larger variance on the high side.

http://www.medtronicdiabetes.com/sites/default/files/library/download-library/user-guides/guardian_real_time_user_guide.pdf